gut(agents): remove model-selection.ts — CLI runtimes own model selection per MBP (#2455)
Delete src/agents/model-selection.ts (170 LOC) per Middleware Boundary
Principle. Model selection is owned by CLI runtimes (Claude, Gemini, Codex,
OpenCode); middleware only reads the configured default from cfg.agents.defaults.

Live config-reading utilities relocated to src/agents/provider-utils.ts
(their natural home alongside existing provider normalization + model ref
parsing helpers):
- buildModelAliasIndex
- resolveConfiguredModelRef
- resolveDefaultModelForAgent
- ModelAliasIndex type

Dead stubs eliminated (zero production callers):
- resolveModelRefFromString (only the test-mock consumed it)
- resolveThinkingDefault (always returned undefined; caller block in chat.ts
  was unreachable — removed)
- getModelRefStatus (always returned undefined)
- inferUniqueProviderFromConfiguredModels (always returned undefined; caller
  block in session-utils.ts was unreachable — removed)
- splitTrailingAuthProfile (internal helper; only served resolveModelRefFromString)

14 callers migrated (issue listed 13 + 1 phantom; this PR additionally covered
provider-capabilities.ts, transcript-policy.ts, and get-reply.test-mocks.ts
which were missed by the issue's path-qualified grep):
- Category A (re-export redirects): agent-runner, directive-handling.model-picker,
  session-store, defaults, gateway-cli-backend.live.test, provider-capabilities,
  transcript-policy
- Category B (type relocation): directive-handling
- Category C (live-function callers): status.summary, sticker-cache, session-utils,
  chat (imports dropped + dead thinkingLevel fallback removed)
- Category D (test mocks): agent.test (with modelSelectionModule →
  providerUtilsModule rename), run.test-harness (mock retarget + stale mock
  helpers removed), get-reply.test-mocks (dead vi.mock block removed)

Zombie-import gate updated:
- scripts/check-no-zombie-imports.mjs: removed "agents/model-selection" from
  fallbackDeadModulePatterns + explanatory block
- scripts/data/zombie-import-allowlist.json: -12 entries
- .zombie-import-allowlist-baseline: 23 → 11

Baselines self-healed:
- .fork-boundary-mock-baseline: 133 → 132 (direct: removed vi.mock in
  get-reply.test-mocks.ts)
- .obsolescence-baseline: 134 → 49 (this PR: -1; pre-existing unbaselined
  improvements: -84 — verified via stash-based pre-change gate run)

Sibling precedents: #2414 (Discord model-picker gut), #2449/#2451 (Mattermost
model-picker gut). Surfaced during #2454 H3 restoration; closes #2355 debt.
