+++ src/agents/transcript-policy.ts
@@ -1,4 +1,4 @@
-import { normalizeProviderId } from "./model-selection.js";
+import { normalizeProviderId } from "./provider-utils.js";
 import type { ToolCallIdMode } from "./tool-call-id.js";
 
 /**
